refactor(memory): move daemon startup orchestration into the plugin (runMemoryStartup)
Extract the ~175-line `initializeQdrantAndMemory` closure out of `lifecycle.ts`
into a plugin-owned `runMemoryStartup(config)` (`plugins/defaults/memory/startup.ts`).
`lifecycle.ts` keeps its single fire-and-forget kick at the same point —
`void runMemoryStartup(config).catch(...)`, still after `startRuntimeHttpServer()` —
so boot timing, the detached-startup philosophy, and the daemon-startup ordering
invariant are preserved. The host stops embedding memory's startup sequence and
drops 13 now-unused memory/persistence imports.

The moved body is logic-identical (verified by diff). The only deltas: prettier
collapses three dynamic imports onto one line each (shorter intra-plugin paths),
and the startup log lines now carry the `memory-startup` logger label instead of
`lifecycle` (nothing keys on logger names).

Boundary baseline +7 — the orchestration's host reaches (Qdrant manager/client,
jobs-worker, register-job-handlers, embedding-reconcile, skill-memory-refresh,
config/schema), now tracked on the plugin rather than hidden in lifecycle.ts.

PR 2 of 2 (startup→plugin); follows the relocate in #36652. Zero
@vellumai/plugin-api changes.
