+++ assistant/src/__tests__/plugin-import-boundary-guard.test.ts
@@ -161,6 +161,7 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../config/assistant-feature-flags.js",
     "../../../config/loader.js",
     "../../../config/memory-v3-gate.js",
+    "../../../config/schema.js",
     "../../../config/schemas/memory-v2.js",
     "../../../config/types.js",
     "../../../contacts/guardian-delivery-reader.js",
@@ -169,10 +170,13 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../context/token-estimator.js",
     "../../../daemon/date-context.js",
     "../../../daemon/disk-pressure-background-gate.js",
+    "../../../daemon/embedding-reconcile.js",
     "../../../daemon/pkb-context-tracker.js",
     "../../../daemon/pkb-reminder-builder.js",
+    "../../../daemon/skill-memory-refresh.js",
     "../../../daemon/tool-setup-types.js",
     "../../../daemon/trust-context.js",
+    "../../../jobs/register-job-handlers.js",
     "../../../permissions/types.js",
     "../../../persistence/checkpoints.js",
     "../../../persistence/conversation-crud.js",
@@ -180,7 +184,10 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../persistence/embeddings/embedding-backend.js",
     "../../../persistence/embeddings/embedding-runtime-manager.js",
     "../../../persistence/embeddings/embedding-types.js",
+    "../../../persistence/embeddings/qdrant-client.js",
+    "../../../persistence/embeddings/qdrant-manager.js",
     "../../../persistence/jobs-store.js",
+    "../../../persistence/jobs-worker.js",
     "../../../persistence/memory-lifecycle-hooks.js",
     "../../../persistence/message-content.js",
     "../../../persistence/raw-query.js",
